Fix TCP chargen crashes
When clients/tcp closes the connection, it would cause the write
in tiprotd to fail with broken pipe.
To shut down gracefully when stdout or stderr gets an EPIPE I had set
a signal handler for SIGPIPE which Rust programs ignore at startup.
That was now being triggered by connections, and would make tiprotd
gracefully shut down after a single connection of this type!
Therefore ignore that signal.

mio poll() also failed with EINTR at one point, so handle that.
This might be a change in mio 0.7?